app-cloud-config. yml
This app-cloud-config.yml file contains the configuration settings required for deploying the mynetwork-app to a cloud environment or server. It includes essential information for connecting the app with external resources such as databases, SSH keys for secure access, and a link to the cloud configuration repository. This file ensures that the app can be properly configured and managed in a cloud-based infrastructure.

cloud_config_url: The URL pointing to the cloud configuration repository or the location where cloud-specific configurations are stored (e.g., on GitHub).

server: The name of the server or environment where the app will be deployed and run.

ssh_key: The SSH key used to authenticate and securely connect to the server.

username: The username for GitHub or the server access.

database: Database configuration details such as host, user, and password required for app connectivity.


This file is essential for ensuring that the app has the correct configuration settings for deployment in a cloud or server environment, including managing secret keys, server access, and database connections.